major overhaul: extend the concept / change the configuration
pazpar2-access is made to be more generic now, to take into account
configuration options other than the GBV login service. For the time being
the extra is the user’s IP address.

pazpar2-access now supports setting any pazpar2 variable (not just pz:allow).
This way we can manipulate variables based on the user’s IP address, e.g.
to provide links to a local catalogue instead of the union catalogue. (So
far the logic for that was part of the display mechanism, now we can
separate it more cleanly into the data processing by configuring pazpar2
to give the desired data right away.)